fix 2

+++ 0x02-python_async_comprehension/2-measure_runtime.py
@@ -6,7 +6,7 @@ import asyncio
 async_comprehension = __import__("1-async_comprehension").async_comprehension
 
 
-async def measure_runtime() -> int:
+async def measure_runtime() -> float:
     """execute async_comprehension four
     times in parallel using asyncio.gather"""
     start = time.perf_counter()
